WIP better claims

+++ app/api/routers/users.py
@@ -1,5 +1,4 @@
-import logging
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException, status
 from app.dependencies import AsyncSessionDep, AccessTokenClaimsDep
 from app.schemas.users import (
     UserResponse,
@@ -11,14 +10,13 @@ from app.schemas.users import (
 from app.modules.auth import hash_password
 from app.services import user_service
 
-log = logging.getLogger(__name__)
 
 router = APIRouter(prefix="/users")
 
 
 @router.get("", response_model=UsersResponse, response_model_exclude_none=True)
 async def get_users(session: AsyncSessionDep):
-    log.info("Getting all users")
+    """List all users"""
     users = await user_service.get_all_users(sess=session)
     return UsersResponse.from_users(users=users)
 
@@ -31,7 +29,7 @@ async def get_users(session: AsyncSessionDep):
 async def get_user_by_pubid(
     session: AsyncSessionDep, pubid: str, incl_items: bool = False
 ):
-    log.info("Getting user %s", pubid)
+    """Get one user by id"""
     if incl_items:
         user = await user_service.get_user_with_items(sess=session, pubid=pubid)
         return UserWithItemsResponse.from_orm(user)
@@ -43,7 +41,7 @@ async def get_user_by_pubid(
     "", response_model=UserResponse, response_model_exclude_none=True, status_code=201
 )
 async def create_user(session: AsyncSessionDep, payload: CreateUserPayload):
-    log.info("Creating new user")
+    """Create a new user"""
     hashed_password = hash_password(pw=payload.password)
 
     user = await user_service.create_user(
@@ -58,11 +56,14 @@ async def create_user(session: AsyncSessionDep, payload: CreateUserPayload):
 @router.put("/{pubid}", response_model=UserResponse, response_model_exclude_none=True)
 async def update_user(
     session: AsyncSessionDep,
-    _: AccessTokenClaimsDep,
+    claims: AccessTokenClaimsDep,
     pubid: str,
     payload: UpdateUserPayload,
 ):
-    log.info("Updating user %s", pubid)
+    """Update an existing user"""
+    if claims["sub"] != pubid:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     user = await user_service.update_user(
         sess=session, pubid=pubid, fullname=payload.fullname
     )
@@ -70,6 +71,11 @@ async def update_user(
 
 
 @router.delete("/{pubid}", response_model_exclude_none=True)
-async def delete_user(session: AsyncSessionDep, _: AccessTokenClaimsDep, pubid: str):
-    log.info("Deleting user %s", pubid)
+async def delete_user(
+    session: AsyncSessionDep, claims: AccessTokenClaimsDep, pubid: str
+):
+    """Delete an existing user"""
+    if claims["sub"] != pubid:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     await user_service.delete_user(sess=session, pubid=pubid)
